Remove DiffEqCallbacks from src
The only place where DiffEqCallbacks was used is to use SavedValues, but
SavedValues is only used in NonInterplatingSavingCallback. Here, I
copied over the definition of SavedValues from DiffEqCallbacks and
removed the direct dependency.

There is still some dependency in the tests, but that can be removed too
if a small amount of work.