Restrict SciMLBase compat to <2.68.0
SciMLBase.remake is called twice during the cache initialization
for a problem with the multirate alg.

SciMLBase changed the way remake(::AbstractODEProblem,...)
works in 2.68.0. Previously, if the f kwarg was an AbstractODEFunction,
the problem was remade with f directly as its function.
After 2.68.0, remake(::AbstractODEProblem) calls a new remake method with
f and prob.f as arguments. This new function returns an
AbstarctSciMLFunction of the same type and properties as prob.f.